+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [3.1.3] - 2024-10-26
+### Added
+- Added passing `original_exception` to fault `!` calls
 ### Changed
 - Move monotonic clock to util
+- Changed `metadata` to a keyword arg on fault `!` calls
 ### Removed
 - Removed unnecessary `Lite::Command::` prefixes